+++ CHANGELOG.md
@@ -507,6 +507,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2942](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2942))
 - `opentelemetry-instrumentation-click`: new instrumentation to trace click commands
   ([#2994](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2994))
+- `opentelemetry-instrumentation-logging` Removed unnecessary calls in logging factory
+  ([#2718](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2718))
+- `opentelemetry-instrumentation-logging` Added the ability to set custom log field names
+  ([#2718](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2718))
 
 ### Fixed